Updated testExec: LPVOID to String
testExec now gets the base address of the kernel & the HEVD driver.
The LPVOID pointer to HEVD is also converted to a string. The HEVD
base address can be used to get the address of any function in the HEVD
driver. A HEVD function address could be added to the shellcode to
return to HEVD!IrpDeviceIoCtlHandler.